+++ practico_08/autogestion_alumnos/templates/login.html
@@ -113,7 +113,7 @@
         </form>
         <div style="width: 700px; margin-top:30px; display: inline-block; text-align: left">
             <a class="signUp" href="/sign_up">Sign Up</a>
-            <a class="lostYourPassword" href="/recover_password">Lost your password?</a>
+            <a class="lostYourPassword" href="/">Lost your password?</a>
         </div>
     </div>
 </div>
