+++ practico_08/run.py
@@ -1,0 +1,5 @@
+from autogestion_alumnos import app, db
+
+if __name__ == '__main__':
+    db.create_all()
+    app.run(port=3000, debug=True)
